criando serviço de gereção de termos

+++ Api/Api/Domain/Interfaces/IEquipamentoRepository.cs
@@ -5,7 +5,7 @@ namespace Api.Domain.Interfaces
     public interface IEquipamentoRepository : IBaseRepository
     {
         Task<IEnumerable<EquipamentoModel>> GetEquipamentosAsync();
-        Task<EquipamentoModel> GetEquipamentoByIdAsync(int id);
+        Task<EquipamentoModel> GetEquipamentoByIdAsync(long id);
 
     }
 }
